+++ plugins/bulk_upload/xml/batch/BulkUploadEngineXml.php
@@ -753,13 +753,20 @@ class BulkUploadEngineXml extends KBulkUploadEngine
 		$requestResults = KBatchBase::$kClient->doMultiRequest();
 
 		$nonCriticalErrors = '';
+		$newThumbAssetsIds = array();
 		foreach($requestResults as $requestResult)
 		{
 			if (is_array($requestResult) && isset($requestResult['code']))
 				$nonCriticalErrors .= $requestResult['message']."\n";
 			if ($requestResult instanceof Exception)
 				$nonCriticalErrors .= $requestResult->getMessage()."\n";
+			if ($requestResult instanceof KalturaThumbAsset)
+				$newThumbAssetsIds[] = $requestResult->id;
 		}
+
+		//delete old thumbnails if necessary
+		if(!$resource && !$keepManualThumbnails  && count($newThumbAssetsIds) > 0)
+			$nonCriticalErrors .= $this->removeOldThumbnails($updatedEntryId, $newThumbAssetsIds);
 
 
 		//update is after add content since in case entry replacement we want the duration to be set on the new entry.
@@ -768,6 +775,31 @@ class BulkUploadEngineXml extends KBulkUploadEngine
 		KBatchBase::unimpersonate();
 		
 		return array($updatedEntry, $nonCriticalErrors);
+	}
+
+	protected function removeOldThumbnails($updatedEntryId, $newThumbAssetsIds)
+	{
+		$filter = new KalturaAssetFilter();
+		$filter->entryIdEqual = $updatedEntryId;
+		$thumbList = KBatchBase::$kClient->thumbAsset->listAction($filter);
+
+		KBatchBase::$kClient->startMultiRequest();
+		foreach($thumbList->objects as $thumbAsset)
+		{
+			if(!in_array($thumbAsset->id, $newThumbAssetsIds))
+				KBatchBase::$kClient->thumbAsset->delete($thumbAsset->id);
+		}
+		$requestResults = KBatchBase::$kClient->doMultiRequest();
+
+		$nonCriticalErrors = '';
+		foreach($requestResults as $requestResult)
+		{
+			if (is_array($requestResult) && isset($requestResult['code']))
+				$nonCriticalErrors .= $requestResult['message']."\n";
+			if ($requestResult instanceof Exception)
+				$nonCriticalErrors .= $requestResult->getMessage()."\n";
+		}
+		return $nonCriticalErrors;
 	}
 
 	/**